Gitlab_Grafana_dashboard/requirement.txt
python-gitlab==3.1.1
prometheus-client==0.13.1